Fix: login, singnup email duplicate check
